Refactor/#56: custom hook 분리
